+++ app/services/flow/flows/provider_capital.rb
@@ -16,7 +16,6 @@ module Flow
         other_assets: Steps::ProviderCapital::OtherAssetsStep,
         restrictions: Steps::ProviderCapital::RestrictionsStep,
         policy_disregards: Steps::ProviderCapital::PolicyDisregardsStep,
-        capital_disregards_discretionary: Steps::ProviderCapital::CapitalDisregardsDiscretionaryStep,
         check_passported_answers: Steps::ProviderCapital::CheckPassportedAnswersStep,
         check_capital_answers: Steps::ProviderCapital::CheckCapitalAnswersStep,
         capital_assessment_results: Steps::ProviderCapital::CapitalAssessmentResultsStep,
