+++ app/services/flow/flows/provider_capital_disregards.rb
@@ -1,0 +1,9 @@
+module Flow
+  module Flows
+    class ProviderCapitalDisregards < FlowSteps
+      STEPS = {
+        capital_disregards_discretionary: Steps::ProviderCapitalDisregards::DiscretionaryStep,
+      }.freeze
+    end
+  end
+end
