+++ app/services/flow/steps/provider_capital_disregards/discretionary_step.rb
@@ -1,7 +1,7 @@
 module Flow
   module Steps
-    module ProviderCapital
-      CapitalDisregardsDiscretionaryStep = Step.new(
+    module ProviderCapitalDisregards
+      DiscretionaryStep = Step.new(
         path: ->(application) { Steps.urls.providers_legal_aid_application_means_capital_disregards_discretionary_path(application) },
         forward: ->(application) { application.passported? ? :check_passported_answers : :check_capital_answers },
         check_answers: ->(application) { application.provider_checking_or_checked_citizens_means_answers? ? :check_capital_answers : :check_passported_answers },
